feature : restructure GenericTable component and implement new Pagination structure with sub-components for improved modularity,
