+++ src/components/common/Pagination/index.ts
@@ -1,0 +1,5 @@
+export { default as Pagination } from './Pagination';
+export { default as PaginationButton } from './PaginationButton';
+export { default as PaginationInfo } from './PaginationInfo';
+export { default as PaginationNavigation } from './PaginationNavigation';
+export { default as PaginationPageNumbers } from './PaginationPageNumbers';
