+++ src/components/projects/MemberAvatar.tsx
@@ -1,56 +1,13 @@
 import { JSX } from 'react';
 
 import TooltipWrapper from '#/components/ui/TooltipWrapper';
+import { getSizeClasses } from '#/lib/utils';
+import { MemberAvatarProps } from '#/types';
 import { cn } from '#/utils';
-
-/**
- * Member role type
- */
-type MemberRole = 'admin' | 'editor' | 'viewer';
-
-/**
- * Project member interface
- */
-interface ProjectMember {
-  id: string;
-  name: string;
-  email: string;
-  avatar?: string;
-  role: MemberRole;
-}
-
-/**
- * Props for MemberAvatar component
- */
-interface MemberAvatarProps {
-  member: ProjectMember;
-  index: number;
-  showTooltip?: boolean;
-  size?: 'sm' | 'md' | 'lg';
-}
 
 /**
  * Get size classes
  */
-const getSizeClasses = (size: 'sm' | 'md' | 'lg'): { avatar: string; role: string } => {
-  switch (size) {
-    case 'sm':
-      return {
-        avatar: 'w-7 h-7 text-xs',
-        role: 'w-4 h-4 text-xs',
-      };
-    case 'lg':
-      return {
-        avatar: 'w-11 h-11 text-sm',
-        role: 'w-6 h-6 text-sm',
-      };
-    default: // md
-      return {
-        avatar: 'w-9 h-9 text-xs',
-        role: 'w-5 h-5 text-xs',
-      };
-  }
-};
 
 /**
  * Member avatar component with role indicator and tooltip
