+++ src/types/genericTable/index.ts
@@ -1,0 +1,1 @@
+export type { Column, TableSection, GenericTableProps } from './genericTable.types';
