+++ src/types/index.ts
@@ -14,3 +14,5 @@ export type * from './common/error.types';
 // Export home types
 export type * from './home/api.types';
 export type * from './home/post.types';
+export type * from './projects';
+export type * from './genericTable';
