+++ src/types/projects/index.ts
@@ -1,0 +1,8 @@
+// Core project types
+export type { ProjectStatus, MemberRole, ProjectMember, ProjectOwner, Project } from './project.types';
+
+// Component-specific types
+export type { ProjectCardProps } from './projectCard.types';
+export type { MemberAvatarProps } from './memberAvatar.types';
+export type { ProjectsTableProps } from './projectsTable.types';
+export type { ProjectStatusColumnsProps } from './projectStatusColumns.types';
