+++ src/types/projects/memberAvatar.types.ts
@@ -1,0 +1,11 @@
+import { ProjectMember } from './project.types';
+
+/**
+ * Props for MemberAvatar component
+ */
+export interface MemberAvatarProps {
+  member: ProjectMember;
+  index: number;
+  showTooltip?: boolean;
+  size?: 'sm' | 'md' | 'lg';
+}
