+++ src/types/projects/projectCard.types.ts
@@ -1,0 +1,10 @@
+import { Project } from './project.types';
+
+/**
+ * Props for ProjectCard component
+ */
+export interface ProjectCardProps {
+  project: Project;
+  onClick?: (project: Project) => void;
+  className?: string;
+}
