+++ src/types/projects/projectsTable.types.ts
@@ -1,0 +1,11 @@
+import { Project } from './project.types';
+
+/**
+ * Props for ProjectsTable component
+ */
+export interface ProjectsTableProps {
+  projects: Project[];
+  onProjectClick: (project: Project) => void;
+  loading?: boolean;
+  className?: string;
+}
